refactor(llm): split generate_litellm_config into litellm_config.py (B7) (#557)
